requirements.txt
newsapi-python==0.2.6
requests==2.23.0
urllib3==1.25.8